move jobtemplate test to its own file
